move powers_of_tau to commitments
