--- include/nil/crypto3/zk/snark/systems/ppzksnark/r1cs_gg_ppzksnark/powers_of_tau/public_key.hpp
@@ -1,48 +1,0 @@
-//---------------------------------------------------------------------------//
-// Copyright (c) 2022 Noam Y <@NoamDev>
-//
-// MIT License
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-//---------------------------------------------------------------------------//
-
-#ifndef CRYPTO3_R1CS_POWERS_OF_TAU_PUBLIC_KEY_HPP
-#define CRYPTO3_R1CS_POWERS_OF_TAU_PUBLIC_KEY_HPP
-
-#include <nil/crypto3/zk/snark/systems/ppzksnark/r1cs_gg_ppzksnark/proof_of_knowledge.hpp>
-
-namespace nil {
-    namespace crypto3 {
-        namespace zk {
-            namespace snark {
-                template<typename CurveType>
-                struct powers_of_tau_public_key {
-                    typedef CurveType curve_type;
-                    typedef proof_of_knowledge<CurveType> pok_type;
-
-                    pok_type tau_pok;
-                    pok_type alpha_pok;
-                    pok_type beta_pok;
-                };
-            }   // snarks
-        }   // zk
-    }   // crypto3
-}   // nil
-
-#endif  // CRYPTO3_R1CS_POWERS_OF_TAU_PUBLIC_KEY_HPP
